requirements.txt
crewai==0.51.1
crewai-tools==0.8.3
atlassian-python-api==3.9.0
python-dotenv==1.0.1
openai==1.51.2  # Para LLM